Fixed problem with Get-ADUser Where statement
There was a problem where the where statement inside of the Get-ADUser command stopped working and would not pull any users. I edited to remove the Threshold and Now variables and just have it coded into the where statement to work. I also added some additional information in the Notes field.